backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0
alembic>=1.13.0
aiosqlite>=0.19.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
openai>=1.12.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.0
pymupdf>=1.23.0
python-pptx>=0.6.21
sentence-transformers>=2.2.0
httpx>=0.26.0
